mix des deux libfts

+++ ft_memchr.c
@@ -6,16 +6,16 @@
 /*   By: Ben <Ben@student.42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/05 16:39:22 by bedesmet          #+#    #+#             */
-/*   Updated: 2021/12/02 00:15:57 by Ben              ###   ########.fr       */
+/*   Updated: 2021/12/07 10:44:50 by Ben              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memchr(const void *s, int c, unsigned int n)
+void	*ft_memchr(const void *s, int c, size_t n)
 {
-	char			*tmp;
-	unsigned int	i;
+	char	*tmp;
+	size_t	i;
 
 	if (n <= 0)
 		return (0);
